refactor: share dendrite handlers
